microservices-auth-project/user-service/requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
requests==2.31.0
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
